+++ apis_v1/urls.py
@@ -7,10 +7,11 @@ This is called from config/urls.py like this:
 """
 
 from django.conf import settings
-from django.conf.urls.static import static
 from django.conf.urls import url
-from . import views
-from . import views_docs
+from django.conf.urls.static import static
+
+from apis_v1.views import views_docs, views_ballot, views_candidate, views_donation, views_facebook, views_friend, \
+    views_misc, views_organization, views_position, views_twitter, views_voter
 from ballot.views_admin import ballot_items_sync_out_view, ballot_returned_sync_out_view
 from candidate.views_admin import candidates_sync_out_view
 from election.views_admin import elections_sync_out_view
@@ -21,136 +22,152 @@ from polling_location.views_admin import polling_locations_sync_out_view
 from position.views_admin import positions_sync_out_view
 from voter_guide.views_admin import voter_guides_sync_out_view
 
+
 urlpatterns = [
     # Actual API Calls
-    url(r'^ballotItemOptionsRetrieve/', views.ballot_item_options_retrieve_view, name='ballotItemOptionsRetrieveView'),
-    url(r'^ballotItemRetrieve/', views.ballot_item_retrieve_view, name='ballotItemRetrieveView'),
+    url(r'^ballotItemOptionsRetrieve/', views_ballot.ballot_item_options_retrieve_view,
+        name='ballotItemOptionsRetrieveView'),
+    url(r'^ballotItemRetrieve/', views_ballot.ballot_item_retrieve_view, name='ballotItemRetrieveView'),
     url(r'^ballotItemsSyncOut/', ballot_items_sync_out_view, name='ballotItemsSyncOutView'),
     url(r'^ballotReturnedSyncOut/', ballot_returned_sync_out_view, name='ballotReturnedSyncOutView'),
-    url(r'^candidateRetrieve/', views.candidate_retrieve_view, name='candidateRetrieveView'),
-    url(r'^candidatesRetrieve/', views.candidates_retrieve_view, name='candidatesRetrieveView'),
+    url(r'^candidateRetrieve/', views_candidate.candidate_retrieve_view, name='candidateRetrieveView'),
+    url(r'^candidatesRetrieve/', views_candidate.candidates_retrieve_view, name='candidatesRetrieveView'),
     url(r'^candidatesSyncOut/', candidates_sync_out_view, name='candidatesSyncOutView'),
-    url(r'^deviceIdGenerate/$', views.device_id_generate_view, name='deviceIdGenerateView'),
-    url(r'^donationWithStripe', views.donation_with_stripe_view, name='donationWithStripeView'),
+    url(r'^deviceIdGenerate/$', views_misc.device_id_generate_view, name='deviceIdGenerateView'),
+    url(r'^donationWithStripe', views_donation.donation_with_stripe_view, name='donationWithStripeView'),
+    url(r'^stripe/', views_donation.donation_stripe_webhook_view, name='donationStripeWebhookView'),
     url(r'^electionsSyncOut/', elections_sync_out_view, name='electionsSyncOutView'),
-    url(r'^facebookDisconnect/', views.facebook_disconnect_view, name='facebookDisconnectView'),
-    url(r'^facebookFriendsAction/', views.facebook_friends_action_view, name='facebookFriendsActionView'),
+    url(r'^facebookDisconnect/', views_facebook.facebook_disconnect_view, name='facebookDisconnectView'),
+    url(r'^facebookFriendsAction/', views_facebook.facebook_friends_action_view, name='facebookFriendsActionView'),
     url(r'^friendInvitationByEmailSend/',
-        views.friend_invitation_by_email_send_view, name='friendInvitationByEmailSendView'),
+        views_friend.friend_invitation_by_email_send_view, name='friendInvitationByEmailSendView'),
     url(r'^friendInvitationByEmailVerify/',
-        views.friend_invitation_by_email_verify_view, name='friendInvitationByEmailVerifyView'),
+        views_friend.friend_invitation_by_email_verify_view, name='friendInvitationByEmailVerifyView'),
     url(r'^friendInvitationByFacebookSend/',
-        views.friend_invitation_by_facebook_send_view, name='friendInvitationByFacebookSendView'),
-    url (r'^friendInvitationByFacebookVerify/',
-        views.friend_invitation_by_facebook_verify_view, name='friendInvitationByFacebookVerifyView'),
+        views_friend.friend_invitation_by_facebook_send_view, name='friendInvitationByFacebookSendView'),
+    url(r'^friendInvitationByFacebookVerify/',
+        views_friend.friend_invitation_by_facebook_verify_view, name='friendInvitationByFacebookVerifyView'),
     url(r'^friendInvitationByWeVoteIdSend/',
-        views.friend_invitation_by_we_vote_id_send_view, name='friendInvitationByWeVoteIdSendView'),
-    url(r'^friendInviteResponse/', views.friend_invite_response_view, name='friendInviteResponseView'),
-    url(r'^friendList/', views.friend_list_view, name='friendListView'),
-    url(r'^measureRetrieve/', views.measure_retrieve_view, name='measureRetrieveView'),
+        views_friend.friend_invitation_by_we_vote_id_send_view, name='friendInvitationByWeVoteIdSendView'),
+    url(r'^friendInviteResponse/', views_friend.friend_invite_response_view, name='friendInviteResponseView'),
+    url(r'^friendList/', views_friend.friend_list_view, name='friendListView'),
+    url(r'^measureRetrieve/', views_misc.measure_retrieve_view, name='measureRetrieveView'),
     url(r'^measuresSyncOut/', measures_sync_out_view, name='measuresSyncOutView'),
-    url(r'^officeRetrieve/', views.office_retrieve_view, name='officeRetrieveView'),
+    url(r'^officeRetrieve/', views_misc.office_retrieve_view, name='officeRetrieveView'),
     url(r'^officesSyncOut/', offices_sync_out_view, name='officesSyncOutView'),
-    url(r'^organizationCount/', views.organization_count_view, name='organizationCountView'),
-    url(r'^organizationFollow/', views.organization_follow_api_view, name='organizationFollowView'),
-    url(r'^organizationFollowIgnore/', views.organization_follow_ignore_api_view, name='organizationFollowIgnoreView'),
-    url(r'^organizationRetrieve/', views.organization_retrieve_view, name='organizationRetrieveView'),
-    url(r'^organizationSave/', views.organization_save_view, name='organizationSaveView'),
-    url(r'^organizationSearch/', views.organization_search_view, name='organizationSearchView'),
+    url(r'^organizationCount/', views_organization.organization_count_view, name='organizationCountView'),
+    url(r'^organizationFollow/', views_organization.organization_follow_api_view, name='organizationFollowView'),
+    url(r'^organizationFollowIgnore/', views_organization.organization_follow_ignore_api_view,
+        name='organizationFollowIgnoreView'),
+    url(r'^organizationRetrieve/', views_organization.organization_retrieve_view, name='organizationRetrieveView'),
+    url(r'^organizationSave/', views_organization.organization_save_view, name='organizationSaveView'),
+    url(r'^organizationSearch/', views_organization.organization_search_view, name='organizationSearchView'),
     url(r'^organizationsSyncOut/', organizations_sync_out_view, name='organizationsSyncOutView'),
     url(r'^organizationStopFollowing/',
-        views.organization_stop_following_api_view, name='organizationStopFollowingView'),
+        views_organization.organization_stop_following_api_view, name='organizationStopFollowingView'),
     url(r'^organizationsFollowedRetrieve/',
-        views.organizations_followed_retrieve_api_view, name='organizationsFollowedRetrieveView'),
+        views_organization.organizations_followed_retrieve_api_view, name='organizationsFollowedRetrieveView'),
     url(r'^organizationSuggestionTasks/',
-        views.organization_suggestion_tasks_view, name='organizationSuggestionTasksView'),
+        views_organization.organization_suggestion_tasks_view, name='organizationSuggestionTasksView'),
     url(r'^pollingLocationsSyncOut/', polling_locations_sync_out_view, name='pollingLocationsSyncOutView'),
     url(r'^positionsCountForAllBallotItems/',
-        views.positions_count_for_all_ballot_items_view, name='positionsCountForAllBallotItemsView'),
+        views_position.positions_count_for_all_ballot_items_view, name='positionsCountForAllBallotItemsView'),
     url(r'^positionsCountForOneBallotItem/',
-        views.positions_count_for_one_ballot_item_view, name='positionsCountForOneBallotItemView'),
-    url(r'^positionLikeCount/', views.position_like_count_view, name='positionLikeCountView'),
-    url(r'^positionListForBallotItem/', views.position_list_for_ballot_item_view, name='positionListForBallotItemView'),
+        views_position.positions_count_for_one_ballot_item_view, name='positionsCountForOneBallotItemView'),
+    url(r'^positionLikeCount/', views_position.position_like_count_view, name='positionLikeCountView'),
+    url(r'^positionListForBallotItem/', views_position.position_list_for_ballot_item_view,
+        name='positionListForBallotItemView'),
     url(r'^positionListForOpinionMaker/',
-        views.position_list_for_opinion_maker_view, name='positionListForOpinionMakerView'),
+        views_position.position_list_for_opinion_maker_view, name='positionListForOpinionMakerView'),
     url(r'^positionListForVoter/',
-        views.position_list_for_voter_view, name='positionListForVoterView'),
+        views_position.position_list_for_voter_view, name='positionListForVoterView'),
     url(r'^positionOpposeCountForBallotItem/',
-        views.position_oppose_count_for_ballot_item_view, name='positionOpposeCountForBallotItemView'),
+        views_position.position_oppose_count_for_ballot_item_view, name='positionOpposeCountForBallotItemView'),
     url(r'^positionPublicOpposeCountForBallotItem/',
-        views.position_public_oppose_count_for_ballot_item_view, name='positionPublicOpposeCountForBallotItemView'),
+        views_position.position_public_oppose_count_for_ballot_item_view,
+        name='positionPublicOpposeCountForBallotItemView'),
     url(r'^positionPublicSupportCountForBallotItem/',
-        views.position_public_support_count_for_ballot_item_view, name='positionPublicSupportCountForBallotItemView'),
-    url(r'^positionRetrieve/', views.position_retrieve_view, name='positionRetrieveView'),
-    url(r'^positionSave/', views.position_save_view, name='positionSaveView'),
+        views_position.position_public_support_count_for_ballot_item_view,
+        name='positionPublicSupportCountForBallotItemView'),
+    url(r'^positionRetrieve/', views_position.position_retrieve_view, name='positionRetrieveView'),
+    url(r'^positionSave/', views_position.position_save_view, name='positionSaveView'),
     url(r'^positionsSyncOut/', positions_sync_out_view, name='positionsSyncOutView'),
     url(r'^positionSupportCountForBallotItem/',
-        views.position_support_count_for_ballot_item_view, name='positionSupportCountForBallotItemView'),
-    url(r'^quickInfoRetrieve/', views.quick_info_retrieve_view, name='quickInfoRetrieveView'),
-    url(r'^searchAll/', views.search_all_view, name='searchAllView'),
-    url(r'^twitterIdentityRetrieve/', views.twitter_identity_retrieve_view, name='twitterIdentityRetrieveView'),
+        views_position.position_support_count_for_ballot_item_view, name='positionSupportCountForBallotItemView'),
+    url(r'^quickInfoRetrieve/', views_misc.quick_info_retrieve_view, name='quickInfoRetrieveView'),
+    url(r'^searchAll/', views_misc.search_all_view, name='searchAllView'),
+    url(r'^twitterIdentityRetrieve/', views_twitter.twitter_identity_retrieve_view, name='twitterIdentityRetrieveView'),
     url(r'^twitterSignInRequestAccessToken/',
-        views.twitter_sign_in_request_access_token_view, name='twitterSignInRequestAccessTokenView'),
+        views_twitter.twitter_sign_in_request_access_token_view, name='twitterSignInRequestAccessTokenView'),
     url(r'^twitterSignInRequestVoterInfo/',
-        views.twitter_sign_in_request_voter_info_view, name='twitterSignInRequestVoterInfoView'),
-    url(r'^twitterSignInStart/', views.twitter_sign_in_start_view, name='twitterSignInStartView'),
-    url(r'^twitterSignInRetrieve/', views.twitter_sign_in_retrieve_view, name='twitterSignInRetrieveView'),
-    url(r'^voterAddressRetrieve/', views.voter_address_retrieve_view, name='voterAddressRetrieveView'),
-    url(r'^voterAddressSave/', views.voter_address_save_view, name='voterAddressSaveView'),
-    url(r'^voterAllPositionsRetrieve/', views.voter_all_positions_retrieve_view, name='voterAllPositionsRetrieveView'),
+        views_twitter.twitter_sign_in_request_voter_info_view, name='twitterSignInRequestVoterInfoView'),
+    url(r'^twitterSignInStart/', views_twitter.twitter_sign_in_start_view, name='twitterSignInStartView'),
+    url(r'^twitterSignInRetrieve/', views_twitter.twitter_sign_in_retrieve_view, name='twitterSignInRetrieveView'),
+    url(r'^voterAddressRetrieve/', views_voter.voter_address_retrieve_view, name='voterAddressRetrieveView'),
+    url(r'^voterAddressSave/', views_voter.voter_address_save_view, name='voterAddressSaveView'),
+    url(r'^voterAllPositionsRetrieve/', views_voter.voter_all_positions_retrieve_view,
+        name='voterAllPositionsRetrieveView'),
     url(r'^voterAllBookmarksStatusRetrieve/',
-        views.voter_all_bookmarks_status_retrieve_view, name='voterAllBookmarksStatusRetrieveView'),
-    url(r'^voterBallotItemsRetrieve/', views.voter_ballot_items_retrieve_view, name='voterBallotItemsRetrieveView'),
+        views_voter.voter_all_bookmarks_status_retrieve_view, name='voterAllBookmarksStatusRetrieveView'),
+    url(r'^voterBallotItemsRetrieve/', views_voter.voter_ballot_items_retrieve_view,
+        name='voterBallotItemsRetrieveView'),
     url(r'^voterBallotItemsRetrieveFromGoogleCivic/',
-        views.voter_ballot_items_retrieve_from_google_civic_view, name='voterBallotItemsRetrieveFromGoogleCivicView'),
-    url(r'^voterBallotListRetrieve/', views.voter_ballot_list_retrieve_view, name='voterBallotListRetrieveView'),
-    url(r'^voterCount/', views.voter_count_view, name='voterCountView'),
-    url(r'^voterCreate/', views.voter_create_view, name='voterCreateView'),
-    url(r'^voterEmailAddressRetrieve/', views.voter_email_address_retrieve_view, name='voterEmailAddressRetrieveView'),
-    url(r'^voterEmailAddressSave/', views.voter_email_address_save_view, name='voterEmailAddressSaveView'),
-    url(r'^voterEmailAddressSignIn/', views.voter_email_address_sign_in_view, name='voterEmailAddressSignInView'),
-    url(r'^voterEmailAddressVerify/', views.voter_email_address_verify_view, name='voterEmailAddressVerifyView'),
-    url(r'^voterExport/', views.VoterExportView.as_view(), name='voterExportView'),
+        views_voter.voter_ballot_items_retrieve_from_google_civic_view,
+        name='voterBallotItemsRetrieveFromGoogleCivicView'),
+    url(r'^voterBallotListRetrieve/', views_voter.voter_ballot_list_retrieve_view, name='voterBallotListRetrieveView'),
+    url(r'^voterCount/', views_voter.voter_count_view, name='voterCountView'),
+    url(r'^voterCreate/', views_voter.voter_create_view, name='voterCreateView'),
+    url(r'^voterEmailAddressRetrieve/', views_voter.voter_email_address_retrieve_view,
+        name='voterEmailAddressRetrieveView'),
+    url(r'^voterEmailAddressSave/', views_voter.voter_email_address_save_view, name='voterEmailAddressSaveView'),
+    url(r'^voterEmailAddressSignIn/', views_voter.voter_email_address_sign_in_view, name='voterEmailAddressSignInView'),
+    url(r'^voterEmailAddressVerify/', views_voter.voter_email_address_verify_view, name='voterEmailAddressVerifyView'),
+    url(r'^voterExport/', views_voter.VoterExportView.as_view(), name='voterExportView'),
     url(r'^voterFacebookSaveToCurrentAccount/',
-        views.voter_facebook_save_to_current_account_view, name='voterFacebookSaveToCurrentAccountView'),
+        views_facebook.voter_facebook_save_to_current_account_view, name='voterFacebookSaveToCurrentAccountView'),
     url(r'^voterFacebookSignInRetrieve/',
-        views.voter_facebook_sign_in_retrieve_view, name='voterFacebookSignInRetrieveView'),
+        views_voter.voter_facebook_sign_in_retrieve_view, name='voterFacebookSignInRetrieveView'),
     url(r'^voterFacebookSignInSave/',
-        views.voter_facebook_sign_in_save_view, name='voterFacebookSignInSaveView'),
+        views_voter.voter_facebook_sign_in_save_view, name='voterFacebookSignInSaveView'),
     url(r'^voterGuidePossibilityRetrieve/',
-        views.voter_guide_possibility_retrieve_view, name='voterGuidePossibilityRetrieveView'),
-    url(r'^voterGuidePossibilitySave/', views.voter_guide_possibility_save_view, name='voterGuidePossibilitySaveView'),
+        views_voter.voter_guide_possibility_retrieve_view, name='voterGuidePossibilityRetrieveView'),
+    url(r'^voterGuidePossibilitySave/', views_voter.voter_guide_possibility_save_view,
+        name='voterGuidePossibilitySaveView'),
     url(r'^voterGuidesFollowedRetrieve/',
-        views.voter_guides_followed_retrieve_view, name='voterGuidesFollowedRetrieveView'),
+        views_voter.voter_guides_followed_retrieve_view, name='voterGuidesFollowedRetrieveView'),
     url(r'^voterGuidesIgnoredRetrieve/',
-        views.voter_guides_ignored_retrieve_view, name='voterGuidesIgnoredRetrieveView'),
+        views_voter.voter_guides_ignored_retrieve_view, name='voterGuidesIgnoredRetrieveView'),
     url(r'^voterGuidesSyncOut/', voter_guides_sync_out_view, name='voterGuidesSyncOutView'),
     url(r'^voterGuidesToFollowRetrieve/',
-        views.voter_guides_to_follow_retrieve_view, name='voterGuidesToFollowRetrieveView'),
+        views_voter.voter_guides_to_follow_retrieve_view, name='voterGuidesToFollowRetrieveView'),
     url(r'^voterLocationRetrieveFromIP/',
-        views.voter_location_retrieve_from_ip_view, name='voterLocationRetrieveFromIPView'),
-    url(r'^voterMergeTwoAccounts/', views.voter_merge_two_accounts_view, name='voterMergeTwoAccountsView'),
-    url(r'^voterOpposingSave/', views.voter_opposing_save_view, name='voterOpposingSaveView'),
-    url(r'^voterPhotoSave/', views.voter_photo_save_view, name='voterPhotoSaveView'),
-    url(r'^voterPositionCommentSave/', views.voter_position_comment_save_view, name='voterPositionCommentSaveView'),
-    url(r'^voterPositionLikeOffSave/', views.voter_position_like_off_save_view, name='voterPositionLikeOffSaveView'),
-    url(r'^voterPositionLikeOnSave/', views.voter_position_like_on_save_view, name='voterPositionLikeOnSaveView'),
+        views_voter.voter_location_retrieve_from_ip_view, name='voterLocationRetrieveFromIPView'),
+    url(r'^voterMergeTwoAccounts/', views_voter.voter_merge_two_accounts_view, name='voterMergeTwoAccountsView'),
+    url(r'^voterOpposingSave/', views_voter.voter_opposing_save_view, name='voterOpposingSaveView'),
+    url(r'^voterPhotoSave/', views_voter.voter_photo_save_view, name='voterPhotoSaveView'),
+    url(r'^voterPositionCommentSave/', views_voter.voter_position_comment_save_view,
+        name='voterPositionCommentSaveView'),
+    url(r'^voterPositionLikeOffSave/', views_voter.voter_position_like_off_save_view,
+        name='voterPositionLikeOffSaveView'),
+    url(r'^voterPositionLikeOnSave/', views_voter.voter_position_like_on_save_view,
+        name='voterPositionLikeOnSaveView'),
     url(r'^voterPositionLikeStatusRetrieve/',
-        views.voter_position_like_status_retrieve_view, name='voterPositionLikeStatusRetrieveView'),
-    url(r'^voterPositionRetrieve/', views.voter_position_retrieve_view, name='voterPositionRetrieveView'),
+        views_voter.voter_position_like_status_retrieve_view, name='voterPositionLikeStatusRetrieveView'),
+    url(r'^voterPositionRetrieve/', views_voter.voter_position_retrieve_view, name='voterPositionRetrieveView'),
     url(r'^voterPositionVisibilitySave/',
-        views.voter_position_visibility_save_view, name='voterPositionVisibilitySaveView'),
-    url(r'^voterRetrieve/', views.voter_retrieve_view, name='voterRetrieveView'),
-    url(r'^voterSignOut/', views.voter_sign_out_view, name='voterSignOutView'),
-    url(r'^voterBookmarkOffSave/', views.voter_bookmark_off_save_view, name='voterBookmarkOffSaveView'),
-    url(r'^voterBookmarkOnSave/', views.voter_bookmark_on_save_view, name='voterBookmarkOnSaveView'),
-    url(r'^voterBookmarkStatusRetrieve/', views.voter_bookmark_status_retrieve_view, name='voterBookmarkStatusRetrieveView'),
-    url(r'^voterStopOpposingSave/', views.voter_stop_opposing_save_view, name='voterStopOpposingSaveView'),
-    url(r'^voterStopSupportingSave/', views.voter_stop_supporting_save_view, name='voterStopSupportingSaveView'),
-    url(r'^voterSupportingSave/', views.voter_supporting_save_view, name='voterSupportingSaveView'),
+        views_voter.voter_position_visibility_save_view, name='voterPositionVisibilitySaveView'),
+    url(r'^voterRetrieve/', views_voter.voter_retrieve_view, name='voterRetrieveView'),
+    url(r'^voterSignOut/', views_voter.voter_sign_out_view, name='voterSignOutView'),
+    url(r'^voterBookmarkOffSave/', views_voter.voter_bookmark_off_save_view, name='voterBookmarkOffSaveView'),
+    url(r'^voterBookmarkOnSave/', views_voter.voter_bookmark_on_save_view, name='voterBookmarkOnSaveView'),
+    url(r'^voterBookmarkStatusRetrieve/', views_voter.voter_bookmark_status_retrieve_view,
+        name='voterBookmarkStatusRetrieveView'),
+    url(r'^voterStopOpposingSave/', views_voter.voter_stop_opposing_save_view, name='voterStopOpposingSaveView'),
+    url(r'^voterStopSupportingSave/', views_voter.voter_stop_supporting_save_view, name='voterStopSupportingSaveView'),
+    url(r'^voterSupportingSave/', views_voter.voter_supporting_save_view, name='voterSupportingSaveView'),
     url(r'^voterTwitterSaveToCurrentAccount/',
-        views.voter_twitter_save_to_current_account_view, name='voterTwitterSaveToCurrentAccountView'),
-    url(r'^voterUpdate/', views.voter_update_view, name='voterUpdateView'),
+        views_voter.voter_twitter_save_to_current_account_view, name='voterTwitterSaveToCurrentAccountView'),
+    url(r'^voterUpdate/', views_voter.voter_update_view, name='voterUpdateView'),
 
     ##########################
     # API Documentation Views
@@ -174,9 +191,9 @@ urlpatterns = [
         views_docs.friend_invitation_by_email_send_doc_view, name='friendInvitationByEmailSendDocs'),
     url(r'^docs/friendInvitationByEmailVerify/$',
         views_docs.friend_invitation_by_email_verify_doc_view, name='friendInvitationByEmailVerifyDocs'),
-    url (r'^docs/friendInvitationByFacebookSend/$',
-         views_docs.friend_invitation_by_facebook_send_doc_view, name='friendInvitationByFacebookSendDocs'),
-url(r'^docs/friendInvitationByFacebookVerify/$',
+    url(r'^docs/friendInvitationByFacebookSend/$',
+        views_docs.friend_invitation_by_facebook_send_doc_view, name='friendInvitationByFacebookSendDocs'),
+    url(r'^docs/friendInvitationByFacebookVerify/$',
         views_docs.friend_invitation_by_facebook_verify_doc_view, name='friendInvitationByFacebookVerifyDocs'),
     url(r'^docs/friendInvitationByWeVoteIdSend/$',
         views_docs.friend_invitation_by_we_vote_id_send_doc_view, name='friendInvitationByWeVoteIdSendDocs'),
