+++ apis_v1/views/views_docs.py
@@ -1,7 +1,7 @@
-# apis_v1/views_docs.py
+# apis_v1/views/views_docs.py
 # Brought to you by We Vote. Be good.
 # -*- coding: UTF-8 -*-
-from .documentation_source import ballot_item_options_retrieve_doc, ballot_item_retrieve_doc, \
+from apis_v1.documentation_source import ballot_item_options_retrieve_doc, ballot_item_retrieve_doc, \
     ballot_items_sync_out_doc, ballot_returned_sync_out_doc, \
     candidate_retrieve_doc, \
     candidates_retrieve_doc, candidates_sync_out_doc, device_id_generate_doc, donation_with_stripe_doc, \
@@ -37,8 +37,8 @@ from .documentation_source import ballot_item_options_retrieve_doc, ballot_item_
     voter_location_retrieve_from_ip_doc, voter_merge_two_accounts_doc, voter_photo_save_doc, \
     voter_position_like_off_save_doc, voter_position_like_on_save_doc, voter_position_like_status_retrieve_doc, \
     voter_position_comment_save_doc, voter_position_retrieve_doc, voter_position_visibility_save_doc, \
-    voter_opposing_save_doc, voter_retrieve_doc, voter_sign_out_doc, voter_bookmark_off_save_doc, voter_bookmark_on_save_doc, \
-    voter_bookmark_status_retrieve_doc, voter_stop_opposing_save_doc, \
+    voter_opposing_save_doc, voter_retrieve_doc, voter_sign_out_doc, voter_bookmark_off_save_doc, \
+    voter_bookmark_on_save_doc, voter_bookmark_status_retrieve_doc, voter_stop_opposing_save_doc, \
     voter_stop_supporting_save_doc, voter_supporting_save_doc, voter_twitter_save_to_current_account_doc, \
     voter_update_doc
 from config.base import get_environment_variable
@@ -214,7 +214,8 @@ def friend_invitation_by_facebook_send_doc_view(request):
     Show documentation about friendInvitationByFacebookSend
     """
     url_root = WE_VOTE_SERVER_ROOT_URL
-    template_values = friend_invitation_by_facebook_send_doc.friend_invitation_by_facebook_send_doc_template_values(url_root)
+    template_values = friend_invitation_by_facebook_send_doc.friend_invitation_by_facebook_send_doc_template_values(
+        url_root)
     template_values['voter_api_device_id'] = get_voter_api_device_id(request)
     return render(request, 'apis_v1/api_doc_page.html', template_values)
 
@@ -896,7 +897,8 @@ def voter_all_bookmarks_status_retrieve_doc_view(request):
     """
     url_root = WE_VOTE_SERVER_ROOT_URL
 
-    template_values = voter_all_bookmarks_status_retrieve_doc.voter_all_bookmarks_status_retrieve_doc_template_values(url_root)
+    template_values = voter_all_bookmarks_status_retrieve_doc.voter_all_bookmarks_status_retrieve_doc_template_values(
+        url_root)
     template_values['voter_api_device_id'] = get_voter_api_device_id(request)
     return render(request, 'apis_v1/api_doc_page.html', template_values)
